Refactor more unit tests into the galaxy-data package.
