--- packages/data/tests/datatypes
@@ -1,1 +1,0 @@
-../../../test/unit/datatypes
