--- packages/data/tests/test_galaxy_mapping.py
@@ -1,1 +1,0 @@
-../../../test/unit/test_galaxy_mapping.py
