+++ test/unit/data/datatypes/test_datatypes_registry.py
